Extract table properties editing tests.
